# fast-api/gen/src/IO.Swagger/Controllers/CategoriesApi.cs
/*
 * Fast Gaming Community API
 *
 * API serving data gathered and analyzed by our lord and savior Cornix
 *
 * OpenAPI spec version: 0.0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class CategoriesApiController : ControllerBase
    { 
        /// <summary>
        /// Delete a specific category
        /// </summary>
        /// <param name="id">Category Id</param>
        /// <response code="204">No Content</response>
        [HttpDelete]
        [Route("/api/v1/categories/{id}")]
        [ValidateModelState]
        [SwaggerOperation("Delete")]
        public virtual IActionResult Delete([FromRoute][Required]string id)
        { 
            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);

            throw new NotImplementedException();
        }

        /// <summary>
        /// List all categories
        /// </summary>
        /// <response code="200">An array of categories</response>
        [HttpGet]
        [Route("/api/v1/categories")]
        [ValidateModelState]
        [SwaggerOperation("Get")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Category>), description: "An array of categories")]
        public virtual IActionResult Get()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Category>));
            string exampleJson = null;
            exampleJson = "[ {\n  \"buy\" : 5,\n  \"sell\" : 5,\n  \"name\" : \"name\",\n  \"categoryId\" : 0,\n  \"categoryItems\" : [ {\n    \"buy\" : 6,\n    \"sell\" : 1,\n    \"name\" : \"name\",\n    \"id\" : \"id\"\n  }, {\n    \"buy\" : 6,\n    \"sell\" : 1,\n    \"name\" : \"name\",\n    \"id\" : \"id\"\n  } ],\n  \"info\" : \"info\"\n}, {\n  \"buy\" : 5,\n  \"sell\" : 5,\n  \"name\" : \"name\",\n  \"categoryId\" : 0,\n  \"categoryItems\" : [ {\n    \"buy\" : 6,\n    \"sell\" : 1,\n    \"name\" : \"name\",\n    \"id\" : \"id\"\n  }, {\n    \"buy\" : 6,\n    \"sell\" : 1,\n    \"name\" : \"name\",\n    \"id\" : \"id\"\n  } ],\n  \"info\" : \"info\"\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Category>>(exampleJson)
                        : default(List<Category>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// List all items for category with id
        /// </summary>
        /// <param name="id">Category Id</param>
        /// <response code="200">An array of category items</response>
        [HttpGet]
        [Route("/api/v1/categories/{id}/items")]
        [ValidateModelState]
        [SwaggerOperation("GetItemsForCategory")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Item>), description: "An array of category items")]
        public virtual IActionResult GetItemsForCategory([FromRoute][Required]string id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Item>));
            string exampleJson = null;
            exampleJson = "[ {\n  \"buy\" : 6,\n  \"sell\" : 1,\n  \"name\" : \"name\",\n  \"id\" : \"id\"\n}, {\n  \"buy\" : 6,\n  \"sell\" : 1,\n  \"name\" : \"name\",\n  \"id\" : \"id\"\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Item>>(exampleJson)
                        : default(List<Item>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Upsert a category
        /// </summary>
        /// <param name="body">New or updated category</param>
        /// <response code="201">Category</response>
        [HttpPost]
        [Route("/api/v1/categories")]
        [ValidateModelState]
        [SwaggerOperation("Post")]
        [SwaggerResponse(statusCode: 201, type: typeof(Category), description: "Category")]
        public virtual IActionResult Post([FromBody]Category body)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201, default(Category));
            string exampleJson = null;
            exampleJson = "{\n  \"buy\" : 5,\n  \"sell\" : 5,\n  \"name\" : \"name\",\n  \"categoryId\" : 0,\n  \"categoryItems\" : [ {\n    \"buy\" : 6,\n    \"sell\" : 1,\n    \"name\" : \"name\",\n    \"id\" : \"id\"\n  }, {\n    \"buy\" : 6,\n    \"sell\" : 1,\n    \"name\" : \"name\",\n    \"id\" : \"id\"\n  } ],\n  \"info\" : \"info\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Category>(exampleJson)
                        : default(Category);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}