# fast-api/gen/src/IO.Swagger/Models/ContainerItem.cs
/*
 * Fast Gaming Community API
 *
 * API serving data gathered and analyzed by our lord and savior Cornix
 *
 * OpenAPI spec version: 0.0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class ContainerItem : IEquatable<ContainerItem>
    { 
        /// <summary>
        /// Gets or Sets Item
        /// </summary>
        [Required]
        [DataMember(Name="item")]
        public Item Item { get; set; }

        /// <summary>
        /// Gets or Sets Droprate
        /// </summary>
        [Required]
        [DataMember(Name="droprate")]
        public double? Droprate { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class ContainerItem {\n");
            sb.Append("  Item: ").Append(Item).Append("\n");
            sb.Append("  Droprate: ").Append(Droprate).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((ContainerItem)obj);
        }

        /// <summary>
        /// Returns true if ContainerItem instances are equal
        /// </summary>
        /// <param name="other">Instance of ContainerItem to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(ContainerItem other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Item == other.Item ||
                    Item != null &&
                    Item.Equals(other.Item)
                ) && 
                (
                    Droprate == other.Droprate ||
                    Droprate != null &&
                    Droprate.Equals(other.Droprate)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Item != null)
                    hashCode = hashCode * 59 + Item.GetHashCode();
                    if (Droprate != null)
                    hashCode = hashCode * 59 + Droprate.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(ContainerItem left, ContainerItem right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(ContainerItem left, ContainerItem right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}